+++ project/game_components/movement/move_component.h
@@ -30,17 +30,17 @@ struct Move_Component
     bool* reflected_rect_indices;
 
     SDL_FRect**  manipulated_rects;
-    size_t       rect_count;
+    unsigned short       rect_count;
     struct Vec2* offsets;
     struct Vec2  max_offset;
-    size_t*      rects_pt_indices;
+    unsigned short*      rects_pt_indices;
 };
 
 
 /* Predef */
 
 struct Move_Component init_move_component(const struct Path path, const time_span_ms step, bool random_x_reflect, int *const exit_code);
-void                couple_move_component(struct Move_Component *const target, SDL_FRect *const manipulated_rects, const size_t rect_count, struct Vec2 max_offset, bool randomize_positions, int *const exit_code);
+void                couple_move_component(struct Move_Component *const target, SDL_FRect *const manipulated_rects, const unsigned short rect_count, struct Vec2 max_offset, bool randomize_positions, int *const exit_code);
 void move_all_rects     (struct Move_Component *const target);
 void free_move_component(struct Move_Component *const target);
