+++ project/helpers/helpers.c
@@ -2,7 +2,7 @@
 #include "helpers.h"
 
 /* Headers */
-#include <stdio.h>    /// `size_t`.
+#include <stdio.h>    /// `unsigned short`.
 #include <stdlib.h>   /// `free()`.
 #include <windows.h>  /// Console text coloring.
 #include "../debug.h" /// Error message printing.
@@ -10,20 +10,20 @@
 
 /* Predef */
 
-void   free_ptr_arr(void **const target, const size_t len);
-char** read_file_by_line(const char *const path, const size_t target_lines);
+void   free_ptr_arr(void **const target, const unsigned short len);
+char** read_file_by_line(const char *const path, const unsigned short target_lines);
 void   textcolor(const unsigned short int color_code);
 
 
 /* Body */
 
 
-void free_ptr_arr(void **const target, const size_t len)
+void free_ptr_arr(void **const target, const unsigned short len)
 {
     if (target == NULL)
         return;
     
-    for (size_t i = 0; i < len; i++)
+    for (unsigned short i = 0; i < len; i++)
     {
         free(target[i]);
         target[i] = NULL;
@@ -34,7 +34,7 @@ void free_ptr_arr(void **const target, const size_t len)
 }
 
 
-char** read_file_by_line(const char *const path, const size_t target_lines)
+char** read_file_by_line(const char *const path, const unsigned short target_lines)
 {
     FILE* data_file = fopen(path, "r");
     if (data_file == NULL)
@@ -51,13 +51,13 @@ char** read_file_by_line(const char *const path, const size_t target_lines)
         data_file = NULL;
         return NULL;
     }
-    for (size_t i = 0; i < target_lines; i++)
+    for (unsigned short i = 0; i < target_lines; i++)
     {
         result[i] = malloc(64 * sizeof(char));
         if (result[i] == NULL)
         {
             print_error("`read_file_by_line()`: couldnt't allocate memory for result lines");
-            for (size_t j = 0; j < i; j++)
+            for (unsigned short j = 0; j < i; j++)
             {
                 free(result[j]);
                 result[j] = NULL;
@@ -69,12 +69,12 @@ char** read_file_by_line(const char *const path, const size_t target_lines)
         }
     }
 
-    for (size_t i = 0; i < target_lines; i++)
+    for (unsigned short i = 0; i < target_lines; i++)
     {
         if (fgets(result[i], 64, data_file) == NULL)
         {
             print_error("`read_file_by_line()`: premature EOF");
-            for (size_t j = 0; j < target_lines; j++)
+            for (unsigned short j = 0; j < target_lines; j++)
             {
                 free(result[j]);
                 result[j] = NULL;
